+++ changelog.md
@@ -3,6 +3,7 @@
 #### v2.0.0
 
 * `[usage]` Added color tags support for description
+* `[terminal]` Improved reading y/n answers
 * `[usage]` Code refactoring
 
 #### v1.8.3