+++ changelog.md
@@ -3,7 +3,8 @@
 #### v2.0.0
 
 * `[usage]` Added color tags support for description
-* `[terminal]` Improved reading y/n answers
+* `[terminal]` Improved reading y/n answers (_incompatible changes_)
+* `[strutil]` Added method `Fields` for "smart" string splitting
 * `[usage]` Code refactoring
 
 #### v1.8.3